♻️ refactor: params->searchparams로 메인 페이지 리팩터
